list_service/requirements.txt
flask==3.0.3
mongoengine==0.27.0
python-dotenv==1.0.1
requests==2.31.0
pymongo==4.6.1




